Add more test calls
